+++ src/main/java/dt/repository/AnswerRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AnswerRepository extends CrudRepository<Answer, Integer> {
+public interface AnswerRepository extends CrudRepository<Answer, Long> {
     List<Answer> findAllByPatientAndTest(Patient patient, Test test);
 }
