+++ src/main/java/dt/repository/ConnectPointsRepository.java
@@ -3,5 +3,5 @@ package dt.repository;
 import dt.model.ConnectPoint;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ConnectPointsRepository extends CrudRepository<ConnectPoint,Integer> {
+public interface ConnectPointsRepository extends CrudRepository<ConnectPoint,Long> {
 }
