+++ src/main/java/dt/repository/PatientRepository.java
@@ -3,7 +3,7 @@ package dt.repository;
 import dt.model.Patient;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PatientRepository extends CrudRepository<Patient,Integer> {
+public interface PatientRepository extends CrudRepository<Patient,Long> {
 
-    Patient findByUserAccountId(int userId);
+    Patient findByUserAccountId(long userId);
 }
