+++ src/main/java/dt/repository/QuestionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface QuestionRepository extends CrudRepository<Question,Integer> {
+public interface QuestionRepository extends CrudRepository<Question,Long> {
 
 //    public List<Question> findByTestConfigurationId(int TestConfigurationId);
 }
