+++ src/main/java/dt/repository/TestConfigurationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.security.access.prepost.PreAuthorize;
 
 @PreAuthorize("hasRole('ROLE_ADMIN')")
-public interface TestConfigurationRepository extends CrudRepository<TestConfiguration,Integer>{
+public interface TestConfigurationRepository extends CrudRepository<TestConfiguration,Long>{
 }
