+++ src/main/java/dt/repository/TestRepository.java
@@ -5,6 +5,6 @@ import dt.model.Test;
 import dt.model.TestConfiguration;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TestRepository extends CrudRepository<Test,Integer> {
+public interface TestRepository extends CrudRepository<Test,Long> {
     Test findByPatientAndTestConfiguration(Patient patient, TestConfiguration testConfiguration);
 }
