+++ src/main/java/dt/service/QuestionService.java
@@ -2,9 +2,7 @@ package dt.service;
 
 import dt.model.ConnectPoint;
 import dt.model.Question;
-import dt.model.Test;
 import dt.model.TestConfiguration;
-import dt.repository.ConnectPointsRepository;
 import dt.repository.QuestionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -29,12 +27,12 @@ public class QuestionService {
         return questions;
     }
 
-    public void saveQuestion(Question question, int testConfigurationId) {
+    public void saveQuestion(Question question, long testConfigurationId) {
 
         questionRepository.save(question);
     }
 
-    public Question getQuestion(int questionId) {
+    public Question getQuestion(long questionId) {
         return questionRepository.findOne(questionId);
     }
 
@@ -42,11 +40,11 @@ public class QuestionService {
         questionRepository.save(question);
     }
 
-    public void deleteQuestion(int questionId) {
+    public void deleteQuestion(long questionId) {
         questionRepository.delete(questionId);
     }
 
-    public List<Question> getAllQuestionsFromTestConfig(int testConfigId) {
+    public List<Question> getAllQuestionsFromTestConfig(long testConfigId) {
         TestConfiguration testConfiguration = testConfigurationService.getTestConfiguration(testConfigId);
         if(testConfiguration != null) {
             return testConfiguration.getQuestions();
@@ -55,7 +53,7 @@ public class QuestionService {
         }
     }
 
-    public void saveQuestionsToTestConfiguration(int testConfigId, List<Question> questions) {
+    public void saveQuestionsToTestConfiguration(long testConfigId, List<Question> questions) {
         TestConfiguration testConfiguration = testConfigurationService.getTestConfiguration(testConfigId);
         if(testConfiguration != null) {
             questions.forEach(testConfiguration.getQuestions()::add);
@@ -67,14 +65,14 @@ public class QuestionService {
 
     public void updateQuestions(List<Question> questions) {
 
-//        for( int i = 0; i < questions.size(); i++ ) {
+//        for( long i = 0; i < questions.size(); i++ ) {
 //            questionRepository.save(questions.get(i));
 //        }
 
         questionRepository.save(questions);
     }
 
-    public void savePointsToQuestion(int questionId, List<ConnectPoint> points) {
+    public void savePointsToQuestion(long questionId, List<ConnectPoint> points) {
         Question question = questionRepository.findOne(questionId);
 
         for (ConnectPoint point : points) {
